Fixed camera selection and minor tweaks to GUI
-Camera numbers were being arbitrarily chosen so I added logic to
specifically open a particular camera based on its unique DeviceID.
-Added a "verbose" checkbox to Arduino version so command window doesn't
display stream warnings unless you want it to.
-Reworded the GUI portion for Arduino version that's related to "TDT
mode selection" in the TDT version so that it's not as confusing, and
made it so that "block name" is automatically generated based on mouse
ID, current date, and specified session number.