First implementation of 2nd order conditioning
This commit implements a hack for 2nd order conditioning on the Arduino
by using the US as CS2. I also cleaned up the actual Arduino code to fix
some persistent bugs with timing and to make bigger use of functions.

There were no changes to the TDT code. This meant that I had to move
"trial table" related files out of the Shared directory because the
trial table is now different between the Arduino and TDT versions. I
will fix this in a subsequent commit.