Added laser stim during US to main program
No longer as much of a hack, though the program does expect the laser to
turn on and off in a specific order relative to CS and US, namely:

csOn --> laserOn --> usOn --> usOff --> laserOff --> csOff.

If you set up the parameters to violate that order the Arduino will
likely get confused and the timing will be garbage.