Add ability for user to specify camera associated with rig
The additions to InitCam give this function the ability to ask the user
for help in specifying which camera is associated with which rig
(necessary because Matlab no longer has the ability to access the gigE
cameras' serial numbers). First, the code checks whether the camera's
exposure time can help distinguish between the rigs. If .the exposure
time value does not help distinguish between the rigs, the function
shows the user what one camera is streaming and asks the user which rig
the video is from. Using that information and the rig number specified
in calling neuroblinks, InitCam connects to/stays connected to the
appropriate camera. Any changes to Launch and to neuroblinks are just to
allow InitCam to take over in finding the right camera to use for the
behavioral session.

So far, I can't get exposure time (src.ExposureTimeAbs) to be saved
outside of a Matlab instance. So, if I set camera 2's exposure time to
4902 and connect to camera 1 with the first instance of Matlab, InitCam
in the second instance of Matlab will read camera 2's exposure time as
4900 and be unable to automatically connect. I tried commenting out
imaqreset to see if that would help, but it doesn't seem to have made a
difference.