Add Note Window functionality and some cosmetic changes
I created a gui that you can launch and use to take notes about certain
times or trials in the session. It hasn't been added to the MainWindow
yet, but you can call it using NoteWindow. Mostly I made this commit to
save my progress on the gui before I go further with it.

I also made some cosmetic changes to InitCam (got rid of an extra
display line) and Launch (got rid of an input variable that doesn't get
used anymore)