fix high frame rate toggle issue
Added a property to metadata (metadata.cam.rignum) to hold the rig that
the camera is associated with. Now, we can add metadata.cam.rignum to
the high frame rate exposure time value or to the low frame rate
exposure time value in Main Window's High Framerate toggle button
callback function to retain rig ID in the frame rate.

Still working on the new GUI